added stuff

+++ db/seeds.rb
@@ -1,11 +1,15 @@
-require 'faker'
+require 'open-uri'
 
-puts "Cleaning database..."
-Ingredient.destroy_all
-puts "Creating cocktails..."
-10.times do
-   Ingredient.create!(
-      name: Faker::Food.fruits,
-    )
+puts "Destroy ingredients"
+Ingredient.destroy_all if Rails.env.development?
+
+puts "Destroy Cocktails"
+Cocktail.destroy_all if Rails.env.development?
+
+puts "Create ingredients"
+url = "https://www.thecocktaildb.com/api/json/v1/1/list.php?i=list"
+ingredients = JSON.parse(open(url).read)
+ingredients["drinks"].each do |ingredient|
+  i = Ingredient.create(name: ingredient["strIngredient1"])
+  puts "create #{i.name}"
 end
-puts "Finished!"
